feat:step 4.21

+++ part4/controllers/blog.js
@@ -100,11 +100,33 @@ blogsRouter.get('/:id', async(request, response, next) => {
 //     .catch((error) => next(error));
 // })
 blogsRouter.delete('/:id', async(request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (error) {
-    next(error)
+  // try {
+  //   await Blog.findByIdAndRemove(request.params.id)
+  //   response.status(204).end()
+  // } catch (error) {
+  //   next(error)
+  // }
+  // step 4.21
+  const blog = await Blog.findById(request.params.id);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: "token missing or invalid"
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  console.log('blog',blog )
+  console.log('user',user)
+  if (blog?.user.toString() === user?.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({
+      error: "you do not have permission to delete this blog",
+    });
   }
 })
